add ResponseEntity

+++ src/main/java/pl/coderslab/author/AuthorController.java
@@ -1,7 +1,8 @@
 package pl.coderslab.author;
 
-import com.sun.org.glassfish.gmbal.ParameterNames;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -43,7 +44,12 @@ class AuthorController {
     }
 
     @DeleteMapping("/{id}")
-    void  deleteAuthor (@PathVariable long id){
+    ResponseEntity<Author> deleteAuthor(@PathVariable long id) {
+        Author author = authorService.getById(id);
+        if (author == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         authorService.delete(id);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
